% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoipresets.R
\docType{data}
\name{aoipreset_types}
\alias{aoipreset_types}
\alias{aoipreset_idflds}
\alias{aoipreset_idval}
\title{Area-of-interest presets}
\format{
An character vector with the names of area-of-interest presets

Named list with with one element per preset type

Named list of data frames, with with one element per preset type
}
\usage{
data(aoipreset_types)

data(aoipreset_idflds)

data(aoipreset_idval)
}
\description{
Area of interest presets

The field(s) each AOI Preset provides to identify features

Values that can be used to identify features in AOI Preset
}
\section{Functions}{
\itemize{
\item \code{aoipreset_idflds}: AOI Preset id fields

\item \code{aoipreset_idval}: Values that can be used to identify features

}}
\keyword{aoipreset}
