% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_resize.R
\name{bbox_resize}
\alias{bbox_resize}
\title{Resize a bounding box object}
\usage{
bbox_resize(x, scale = NULL, buff = NULL)
}
\arguments{
\item{x}{bbox object}

\item{scale}{scale factor}

\item{buff}{buffer distance}
}
\description{
Resize a bounding box object by a scale factor or buffer distance
}
\details{
You can resize a bounding box by passing a value for \code{scale} or \code{buff}, but not both.

Use \code{scale} to resize the bounding box with a scale factor. Values of \code{scale} < 1 will
result in a smaller bounding box, values > 1 will result in a larger bounding box. The centroid will
remain the same. If you pass two values for \code{scale}, they'll be used to scale the x and y
dimensions respectively

Use \code{buff} to resize the bounding box with a fixed distance. \code{buff} should be in a map units.
Values of \code{buff} < 0 will result in a smaller bounding box. Values > 0 will result in a larger
bounding box. If you pass two values for \code{buff}, they'll be used to buffer the x and y
dimensions respectively
}
