% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_aoipreset_geom.R
\name{ca_aoipreset_geom}
\alias{ca_aoipreset_geom}
\title{Get the geometry of an AOI Preset area}
\usage{
ca_aoipreset_geom(aoipreset, quiet = TRUE)
}
\arguments{
\item{aoipreset}{The name of a AOI preset}

\item{quiet}{Suppress messages}
}
\value{
A simple feature data frame
}
\description{
Get the geometry of an AOI Preset area
}
\details{
This retrieves the geometry (i.e., boundaries) for one of Cal-Adapt's AOI Presets.
If the spatial layer has not already been downloaded, it will be downloaded
(from \url{https://github.com/ucanr-igis/caladaptr/tree/master/aoipreset_geoms}) and saved in
the local cache directory as a GeoPackage. The default local cache directory is buried in
the current user's 'AppData' folder. To put the GeoPackages in an easier-to-find location,
use \code{\link{ca_setcache}}.
}
\seealso{
\code{\link{aoipreset_types}}, \code{\link{ca_getcache}}
}
