% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_apicalls.R
\name{ca_apicalls}
\alias{ca_apicalls}
\title{Creates a tibble of individual API calls}
\usage{
ca_apicalls(x, slug_check = TRUE, date_check = TRUE)
}
\arguments{
\item{x}{A Cal-Adapt API request}

\item{slug_check}{Cross check the slug against the raster series catalog}

\item{date_check}{Cross check the start and end date against the raster series catalog}
}
\description{
Constructs a tibble of individual API calls for an API request
}
