% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseurl.R
\docType{data}
\name{ca_baseurl}
\alias{ca_baseurl}
\title{Cal-Adapt Base URL}
\format{
An character vector
}
\usage{
data(ca_baseurl)
}
\description{
Base URL for all calls to the Cal-Adapt API
}
\keyword{datasets}
