% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_catalog_aoipreset.R
\name{ca_catalog_aoipreset}
\alias{ca_catalog_aoipreset}
\title{Retrieves Cal-Adapt data catalog}
\usage{
ca_catalog_aoipreset(
  aoipreset = ca_aoipreset,
  download = TRUE,
  cache = TRUE,
  save_aoigeom = TRUE,
  max_pages = NULL,
  quiet = FALSE
)
}
\arguments{
\item{aoipreset}{AOI preset(s)}

\item{download}{Download new catalog from Cal-Adapt}

\item{cache}{Use local cache (see details)}

\item{save_aoigeom}{Save the geom for AOI preset}

\item{max_pages}{Maximum number of pages to retrieve}

\item{quiet}{Suppress messages}
}
\value{
A named list of data frames. Each list element corresponds to one of the
AOI Preset types (e.g., counties). The data frame includes the columns which uniquely
identify each feature (e.g., name, fips code). #'
}
\description{
Retrieves Cal-Adapt data catalog
}
\details{
\code{aoipreset} can be one of the values in \code{ca_aoipreset}.

If \code{cache=TRUE} and \code{download=FALSE}, it will look in the cache directory
before using the catalog bundled with \code{caladaptr}.  If \code{download=TRUE} and
\code{cache=TRUE}, the new results will be saved.
}
