% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_catalog_fetch.R
\name{ca_catalog_fetch}
\alias{ca_catalog_fetch}
\title{Fetch a new copy of the Cal-Adapt raster series catalog}
\usage{
ca_catalog_fetch(quiet = FALSE, save_to_cache = TRUE)
}
\arguments{
\item{quiet}{Suppress messages, logical}

\item{save_to_cache}{Save the catalog to the cache directory, logical}
}
\value{
The raster series catalog as a tibble
}
\description{
Fetch a new copy of the Cal-Adapt raster series catalog
}
\details{
This function will download a list of all raster series available through the Cal-Adapt API.
If \code{save_to_cache = TRUE}, the catalog will be saved as a csv file in the caladaptR's cache folder
and used for subsequent calls to \code{\link{ca_catalog_rs}}.

A copy of the raster series catalog is also included with caladaptR. You can run \code{ca_catalog_fetch()}
to update the catalog when new raster series are published on Cal-Adapt. The best way to
find out when new data are published on Cal-Adapt is to subscribe to the Cal-Adapt newsletter.
}
\seealso{
\code{\link{ca_catalog_rs}}, \code{\link{ca_getcache}}, \code{\link{ca_setcache}}
}
