% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_catalog_rs.R
\name{ca_catalog_rs}
\alias{ca_catalog_rs}
\title{Get the Cal-Adapt raster series data catalog}
\usage{
ca_catalog_rs(download = FALSE, cache = download, quiet = FALSE)
}
\arguments{
\item{download}{Download a new copy of the catalog from Cal-Adapt.org}

\item{cache}{Use a local copy}

\item{quiet}{Suppress messages}
}
\value{
A data frame (tibble) with columns of information about the raster series available through
the Cal-Adapt API.
}
\description{
Get the Cal-Adapt raster series data catalog
}
\details{
This retrieves a 'catalog' of the raster series available through the Cal-Adapt API.
Normally, you can use the copy which comes with caladaptr, but to force download a new copy,
use \code{download = TRUE}. This is only needed when new
datasets are added to the API (subscribe to the Cal-Adapt news letter for notifications).
}
\seealso{
\code{\link{ca_getcache}}
}
