% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_catalog_search.R
\name{ca_catalog_search}
\alias{ca_catalog_search}
\title{Search the Cal-Adapt raster series data catalog}
\usage{
ca_catalog_search(x, keep_together = FALSE, quiet = FALSE)
}
\arguments{
\item{x}{text to search for}

\item{keep_together}{treat x as a phrase, logical}

\item{quiet}{suppress messages}
}
\value{
A tibble with information about the found raster series
}
\description{
Search the Cal-Adapt raster series data catalog
}
\details{
This function can be used to search the local copy of the Cal-Adapt raster series
data catalog, and view the properties of the matching results. Searched fields include the
dataset name and slug. If \code{keep_together = TRUE}, the search text will be treated
as a phrase, otherwise the words in x will be searched for separately.
Records have to match all terms to be returned.

For an online search tool, click the 'Filters' button on
\url{https://api.cal-adapt.org/api/series/}.
}
\examples{
\dontrun{
## Search for a slug
ca_catalog_search("pr_day_gridmet")

## Search for keywords
ca_catalog_search("evapotranspiration year")

## Search for phrase
ca_catalog_search("Livneh VIC", keep_together = TRUE)
}

}
\seealso{
\code{\link{ca_catalog_rs}}, \code{\link{ca_catalog_fetch}}
}
