% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_check.R
\name{ca_check}
\alias{ca_check}
\title{Check the properties of a Cal-Adapt values databases}
\usage{
ca_check(x)
}
\arguments{
\item{x}{An object to be checked}
}
\description{
Check the properties of a Cal-Adapt values databases
}
\details{

}
\seealso{
\code{\link{ca_getvals}}
}
