% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_climvar.R
\docType{data}
\name{ca_climvar}
\alias{ca_climvar}
\title{Climate Variables}
\format{
An character vector with three abbreviated names of carbon emissions scenarios
}
\source{
\url{https://berkeley-gif.github.io/caladapt-docs/data-catalog.html#climate-variables}
}
\usage{
data(ca_climvar)
}
\description{
Climate variables available as raster series
}
\details{
The following climate variables are available as raster series thru the Cal-Adapt API.

\code{tasmax}: Maximum Temperature (historical values from UW Hydro and forecast values from LOCA downscaled climate projections)

\code{tasmin}: Minimum Temperature (historical values from UW Hydro and forecast values from LOCA downscaled climate projections)

\code{precip}: Precipitation (historical values from UW Hydro and forecast values from LOCA downscaled climate projections)
}
\keyword{datasets}
