% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_cvar.R
\name{ca_cvar}
\alias{ca_cvar}
\title{Assign climate variable(s) to a Cal-Adapt API call}
\usage{
ca_cvar(x = ca_apireq(), cvar)
}
\arguments{
\item{x}{Cal-Adapt API request}

\item{cvar}{Climate variable}
}
\description{
Specifies climate variable(s) a Cal-Adapt API call should retrieve
}
\details{
For valid options for \code{cvar}, see \code{\link{climvars}}.

Notes:

1) 'climate variables' refers to both the variables returned by global
circulation models (e.g., temperature, precipitation), as well as
variables derived by additional models (e.g., evapo-transpiration)

2) Not all climate variables are available for all climate models,
temporal periods, and date ranges.
}
