% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_dataset.R
\name{ca_dataset}
\alias{ca_dataset}
\title{Assigns a dataset to a Cal-Adapt API call}
\usage{
ca_dataset(x = ca_apireq(), gcm, scenario)
}
\arguments{
\item{x}{Cal-Adapt API request}

\item{gcm}{the name of a supported GCM}

\item{scenario}{the name of an emissions scenario}
}
\description{
Specifies a dataset to a Cal-Adapt API call
}
\details{
For valid options for \code{gcm} and \code{scenario}, see the
built-in constants \code{\link{gcms}} and \code{\link{scenarios}}.
}
