% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_dates.R
\name{ca_dates}
\alias{ca_dates}
\title{Adds a start and end date of a Cal-Adapt API call}
\usage{
ca_dates(x = ca_apireq(), start, end)
}
\arguments{
\item{x}{Cal-Adapt API request}

\item{start}{start date entered as a character \emph{yyyy-mm-dd} or Date object}

\item{end}{end date entered as a character \emph{yyyy-mm-dd} or Date object}
}
\description{
Specifies the start and end date of a Cal-Adapt API call
}
