% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_db_indices.R
\name{ca_db_indices}
\alias{ca_db_indices}
\title{Add or delete indices}
\usage{
ca_db_indices(x, tbl, idx_fld_add = NULL, idx_fld_del = NULL, quiet = FALSE)
}
\arguments{
\item{x}{Either a remote tibble or a SQLite database file name}

\item{tbl}{The name of a table in the SQLite database}

\item{idx_fld_add}{Fields in tbl to create an index for}

\item{idx_fld_del}{Fields in tbl that have indices you'd like to delete}

\item{quiet}{Suppress messages}
}
\description{
Add or delete indices
}
\details{
Database indices improve performance when you filter or sort rows based on field (column),
and/or join tables based on a common field. By default, indices are not created when you download
Cal-Adapt data into a SQLite database with \code{\link{ca_getvals_db}} (because they increase the
size of the SQLite file). You can tell \code{\link{ca_getvals_db}} to create indices with the \code{indices}
argument, or use \code{ca_db_indices} to create indices after data are downloaded.

\code{x} can be either a remote tibble returned by \code{\link{ca_getvals_db}}, or a SQLite
database file name. \code{tbl} should be the name of a table in the database (i.e., the \code{db_tbl}
argument you passed to \code{ca_getvals_db}. If you're not sure what the table names are in a database,
run \code{\link{ca_db_info}}. Normally you would only add indices to a table that contains values
from Cal-Adapt (there is no need to add indices to lookup tables). You can only add indices for one
table at a time (but \code{idx_fld_add} can contain multiple field names).

Note that \code{ca_db_indices} can only create indices on a single field. To create composite indices
you can run SQL expressions with the DBI package. Indices added by \code{ca_db_indices} will be named
automatically.

For more details, see the vignette on querying large volumes of data:
\code{vignette("big-query", package = "caladaptr")}
}
\seealso{
\code{\link{ca_getvals_db}}, \code{\link{ca_db_info}}
}
