% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_gcm.R
\name{ca_gcm}
\alias{ca_gcm}
\title{Adds GCM(s) to a Cal-Adapt API request}
\usage{
ca_gcm(x = ca_apireq(), gcm)
}
\arguments{
\item{x}{Cal-Adapt API request}

\item{gcm}{Global Climate Model abbreviation}
}
\description{
Specifies GCM(s) a Cal-Adapt API call should retrieve
}
\details{
For valid options for \code{gcm}, see \code{\link{gcms}}.
}
