% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_getvals.R
\name{ca_getvals}
\alias{ca_getvals}
\title{Get values from an API request object}
\usage{
ca_getvals(x, quiet = FALSE)
}
\arguments{
\item{x}{A Cal-Adapt API request}

\item{quiet}{Suppress messages}
}
\description{
Get values from an API request object
}
