% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_livneh.R
\name{ca_livneh}
\alias{ca_livneh}
\title{Use Livneh data in a Cal-Adapt API call}
\usage{
ca_livneh(x = ca_apireq(), livneh = TRUE)
}
\arguments{
\item{x}{Cal-Adapt API request}

\item{livneh}{Use Livneh data, logical}
}
\description{
Specify Livneh data should be retrieved in a Cal-Adapt API call
}
\details{

}
\seealso{
\code{\link{ca_catalog_rs}}
}
