% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_loc_pt.R
\name{ca_loc_pt}
\alias{ca_loc_pt}
\title{Assign point location to a Cal-Adapt API request}
\usage{
ca_loc_pt(x = ca_apireq(), coords, id = NULL)
}
\arguments{
\item{x}{A Cal-Adapt API request}

\item{coords}{A two-column matrix or data frame}

\item{id}{Unique id values for each point}
}
\description{
Specifies point location(s) a Cal-Adapt API call should retrieve
}
\details{
\code{coords} should be a two-column matrix or data frame with x (longitude)
and y (latitude) coordinates for the point(s) of interest.

\code{id} should be vector that uniquely identify the points. If omitted, row numbers will be used.
}
\seealso{
\code{\link{ca_apireq}}
}
