% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_options.R
\name{ca_options}
\alias{ca_options}
\title{Add processing options to a Cal-Adapt API call}
\usage{
ca_options(
  x = ca_apireq(),
  spatial_ag = c("none", "mean", "max", "median", "min", "sum")[1],
  temporal_ag = NA
)
}
\arguments{
\item{x}{Cal-Adapt API request}

\item{spatial_ag}{Spatial aggregation function for polygon locs.}

\item{temporal_ag}{List object specifying unit of time and summary function(s) for temporal aggregation. NOT YET SUPPORTED}
}
\description{
Specify processing options for a Cal-Adapt API call
}
\details{
\code{spatial_ag} is the name(s) of summary statistic(s) that will be used when querying
polygon locations to aggregate the values of pixel that fall within the
area of interest. Values can be \code{mean}, \code{max}, \code{median}, \code{min},
and \code{sum}. To get retrieve the individual values for all pixels without a summary, use
the \code{ca_getrst} function.

When querying point locations (e.g., \code{\link{ca_loc_pt}} or \code{ca_loc_zip}, \code{spatial_ag} should be set to
\code{'none'}.

\code{temporal_ag} allows you to apply an \emph{additional} temporal
aggregation function, on top of any temporal aggregation the data are already summarized by
(e.g., month or year). When querying climate layers that are already temporally aggregated,
the unit of temporal aggregation must be a larger unit of time (e.g., you can't pull down annual average
layers and then try to aggregate them by month).
}
