% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_period.R
\name{ca_period}
\alias{ca_period}
\title{Assign temporal aggregation period to a Cal-Adapt API call}
\usage{
ca_period(x = ca_apireq(), period)
}
\arguments{
\item{x}{Cal-Adapt API request}

\item{period}{Period of temporal aggregation}
}
\description{
Specifies the temporal aggregation periond a Cal-Adapt API call should retrieve
}
\details{
For valid options for \code{period}, run \code{\link{periods}}.

Notes:
}
