% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_preflight.R
\name{ca_preflight}
\alias{ca_preflight}
\title{Run checks on an API request object}
\usage{
ca_preflight(
  x,
  slug_check = TRUE,
  date_check = TRUE,
  loc_check = TRUE,
  units_check = TRUE,
  ignore_spag = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{x}{A Cal-Adapt API request}

\item{slug_check}{Cross check the slug against the raster series catalog}

\item{date_check}{Cross check the start and end date against the raster series catalog}

\item{loc_check}{Check to make sure the location is within the Cal-Adapt coverage area}

\item{units_check}{Check for consistent units}

\item{ignore_spag}{Ignore the spatial aggregation option}

\item{quiet}{Suppress messages}
}
\description{
Run checks on an API request object
}
\details{
This function checks an API request for errors.

If a slug is not found, the variable may not be available for the temporal aggregation period.
See \code{\link{ca_catalog_rs}}.
}
