% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_db_read.R
\name{ca_read_db}
\alias{ca_read_db}
\title{Load a Cal-Adapt SQLite database into R}
\usage{
ca_read_db(
  x,
  val_tbl = NULL,
  join_lookup_tbls = TRUE,
  all_tables = FALSE,
  exclude_hash_tables = TRUE
)
}
\description{
`r lifecycle::badge(deprecated)`
`ca_read_db()` was renamed to `ca_db_read()` for a more consistent API.
}
\keyword{internal}
