% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_stars_read.R
\name{ca_read_stars}
\alias{ca_read_stars}
\title{Read downloaded TIF files into R}
\usage{
ca_read_stars(x, read_sidecar = TRUE, proxy = FALSE)
}
\description{
`r lifecycle::badge(deprecated)`

`ca_read_stars()` was renamed to `ca_stars_read()` for a more consistent API.
}
\seealso{
\code{\link{ca_stars_read}}
}
\keyword{internal}
