% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_resp_check.R
\name{ca_resp_check}
\alias{ca_resp_check}
\title{Check the response to an API call}
\usage{
ca_resp_check(resp, task_msg)
}
\arguments{
\item{resp}{A response object}

\item{task_msg}{A message describing the API task}
}
\description{
Check the response to an API call
}
