% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_scenario.R
\docType{data}
\name{ca_scenario}
\alias{ca_scenario}
\title{Scenarios}
\format{
An character vector with three abbreviated names of carbon emissions scenarios
}
\source{
\url{https://berkeley-gif.github.io/caladapt-docs/data-catalog.html#scenarios}
}
\usage{
data(ca_scenario)
}
\description{
Carbon emission scenarios available through the Cal-Adapt API
}
\details{
The following carbon emission scenarios are available thru the Cal-Adapt API.

\code{rcp45}: RCP 4.5 (Emissions peak around 2040, then decline)

\code{rcp85}: RCP 8.5 (Emissions continue to rise strongly through 2050 and plateau)

\code{historical}: Historical
}
\keyword{datasets}
