% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_slug.R
\name{ca_slug}
\alias{ca_slug}
\title{Adds slug(s) to a Cal-Adapt API call}
\usage{
ca_slug(x = ca_apireq(), slug)
}
\arguments{
\item{x}{Cal-Adapt API request}

\item{slug}{Raster series slug(s)}
}
\description{
Specify the raster series slug a Cal-Adapt API call should retrieve
}
\details{
To find valid slugs, see \code{\link{ca_catalog_rs}}.
}
\seealso{
\code{\link{ca_catalog_rs}}
}
