% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_stars_mosaic.R
\name{ca_stars_mosaic}
\alias{ca_stars_mosaic}
\title{Mosaic stars objects into a seamless array for large areas}
\usage{
ca_stars_mosaic(
  stars_lst,
  index_tbl = NULL,
  geom_mask = NULL,
  combine_6d = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{stars_lst}{A list of stars rasters}

\item{index_tbl}{A tibble of metadata for stars_lst}

\item{geom_mask}{A sf or sfc polygon object to crop the mosaiced raster}

\item{combine_6d}{Combine multiple 3D rasters into one 6D raster}

\item{quiet}{Suppress messages}
}
\value{
If \code{combine_6d = FALSE}, a list of 3D rasters will be returned (x, y and date/year). If \code{combine_6d = TRUE}, a 6D stars
raster will be returned (x, y, date/year, scenario, GCM, climate variable).
}
\description{
Mosaic stars objects into a seamless array for large areas
}
\details{
stars_lst is a list of stars objects downloaded by \code{\link{ca_getrst_stars}} and turned into a list by
\code{\link{ca_stars_read}}. Note that both of these functions must be run with `sidecar = TRUE` (the default).

\code{combine_6d = TRUE} will return the mosaic as a six-dimensional raster. This further requires that the raster was
originally downloaded using an API request that specified the dataset by scenario, GCM, and climate variable (i.e., not
by a slug).
}
\examples{
\dontrun{

}
}
\seealso{
\code{\link{ca_getrst_stars}}, \code{\link{ca_stars_read}}, \code{\link{ca_stars_index}}
}
