% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_stars_read.R
\name{ca_stars_read}
\alias{ca_stars_read}
\title{Read tif files from disk}
\usage{
ca_stars_read(x, sidecar = TRUE, proxy = FALSE)
}
\arguments{
\item{x}{File name(s) with path of tif files}

\item{sidecar}{Read sidecar files if they exist, logical}

\item{proxy}{Import the TIF file as a stars proxy}
}
\value{
A list of stars objects
}
\description{
Read tif files from disk
}
\details{
This function can be used to read tif files that were downloaded by \code{\link{ca_getrst_stars}}. It
is a lightweight wrapper around stars::read_stars(), with the ability to read the sidecar
files created by \code{\link{ca_getrst_stars}}. These sidecar files restore all of the attributes of
the array dimensions that are not otherwise preserved by the tif format.

\code{proxy} signifies whether the tif file(s) should be read as stars proxy (e.g., pointer).
This is recommended if the rasters are potentially too large to fit into memory. With stars proxy
objects, rasters will be read into memory only when needed, at the cost of potentially slightly
slower performance. For details see
\href{https://r-spatial.github.io/stars/articles/stars2.html#stars-proxy-objects}{stars documentation}.
}
\seealso{
\code{\link{ca_getrst_stars}}
}
