% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_vals2tbl.R
\name{ca_vals2tbl}
\alias{ca_vals2tbl}
\title{Converts a Cal-Adapt values list to a data frame}
\usage{
ca_vals2tbl(
  x,
  feat_id = TRUE,
  dt = TRUE,
  cvar = TRUE,
  period = TRUE,
  gcm = TRUE,
  scenario = TRUE,
  spatial_ag = TRUE,
  force_day_one = TRUE
)
}
\arguments{
\item{x}{A Cal-Adapt Values Object}

\item{feat_id}{Include feature id as a column}

\item{dt}{Include date}

\item{cvar}{Include climate variable}

\item{period}{Include period}

\item{gcm}{Include gcm}

\item{scenario}{Include scenario}

\item{spatial_ag}{Include spatial aggregation function}

\item{force_day_one}{Convert the dt to an integer value when period = year}
}
\description{
Converts a Cal-Adapt values list to a data frame
}
\details{
If \code{force_day_one = TRUE}, the day value of the dt column will be forced to January 1 when the
queried layer is pre-aggregated by year (i.e., \code{period = 'year'}). This ad-hoc conversion helps provide
consistency in how the 'date' of a temporally aggregated layer is returned by the Cal-Adapt events and
rasters end point. Without this conversion, the 'date' of a year may be returned as December 31.
}
\seealso{
\code{\link{ca_getvals}}
}
