% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_years.R
\name{ca_years}
\alias{ca_years}
\title{Adds the start and end year to a Cal-Adapt API call}
\usage{
ca_years(x = ca_apireq(), start, end)
}
\arguments{
\item{x}{Cal-Adapt API request}

\item{start}{start year}

\item{end}{end year}
}
\description{
Specifies the start and end year of a Cal-Adapt API call
}
