% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvars.R
\docType{data}
\name{cvars}
\alias{cvars}
\title{Climate Variables}
\format{
A character vector with the names of climate variables
}
\source{
\url{https://berkeley-gif.github.io/caladapt-docs/data-catalog.html#climate-variables}
}
\usage{
data(cvars)
}
\description{
Climate variables available as raster series
}
\details{
The following climate variables are available as raster series thru the Cal-Adapt API.

\code{tasmax}: Maximum Temperature (historical values from UW Hydro and forecast values from LOCA downscaled climate projections)

\code{tasmin}: Minimum Temperature (historical values from UW Hydro and forecast values from LOCA downscaled climate projections)

\code{pr}: Precipitation (historical values from UW Hydro and forecast values from LOCA downscaled climate projections)
}
\keyword{datasets}
