% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_apireq_methods.R
\name{format.ca_apireq}
\alias{format.ca_apireq}
\title{Format a ca_apireq object}
\usage{
\method{format}{ca_apireq}(x, ...)
}
\arguments{
\item{x}{Cal-Adapt API request}

\item{...}{Unused}
}
\description{
Format a ca_apireq object
}
