% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_db_info.R
\name{format.ca_db_info}
\alias{format.ca_db_info}
\title{Format a ca_db_info object for printing at the console}
\usage{
\method{format}{ca_db_info}(x, ...)
}
\arguments{
\item{x}{An object of class ca_db_info}

\item{...}{Unused}
}
\description{
Format a ca_db_info object for printing at the console
}
