% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcms.R
\docType{data}
\name{gcms}
\alias{gcms}
\title{Global Climate Models}
\format{
An character vector with the names of 10 GCMs (abbreviated)
}
\source{
\url{https://berkeley-gif.github.io/caladapt-docs/data-catalog.html#global-climate-models-gcm}
}
\usage{
data(gcms)
}
\description{
Global Climate Models available through the Cal-Adapt API
}
\details{
The following GCMs have been selected by California state agencies as priority models for Fourth Assessment Research and are available thru the Cal-Adapt API. The first four have been identified as the priority GCMs representing (see \href{https://www.climatechange.ca.gov/climate_action_team/research.html}{4 priority models})

\code{HadGEM2-ES}: Met Office Hadley Centre and Instituto Nacional de Pesquisas Espaciais

\code{CNRM-CM5}: Centre National de Recherches Météorologiques/ Centre Européen de Recherche et Formation Avancée en Calcul Scientifique

\code{CanESM2}: Canadian Centre for Climate Modelling and Analysis

\code{MIROC5}: Atmosphere and Ocean Research Institute (The University of Tokyo), National Institute for Environmental Studies, and Japan Agency for Marine-Earth Science and Technology

\code{ACCESS1-0}: Commonwealth Scientific and Industrial Research Organization (CSIRO) and Bureau of Meteorology (BOM) Australia

\code{CCSM4}: University of Miami - RSMAS

\code{CESM1-BGC}: Community Earth System Model Contributors

\code{CMCC-CMS}: Centro Euro-Mediterraneo per I Cambiamenti Climatici

\code{GFDL-CM3}: NOAA Geophysical Fluid Dynamics Laboratory

\code{HadGEM2-CC}: Met Office Hadley Centre
}
\keyword{datasets}
