% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periods.R
\docType{data}
\name{periods}
\alias{periods}
\title{Temporal aggregation periods}
\format{
An character vector with three names of temporal aggregation periods
}
\source{
\url{https://berkeley-gif.github.io/caladapt-docs/data-catalog.html#period}
}
\usage{
data(periods)
}
\description{
Temporal aggregation periods for raster series
}
\details{
The following temporal aggregation periods for raster series are available thru the Cal-Adapt API.

\code{day}: Daily values
\code{month}: Monthly summary statistic
\code{year}: Annual summary statistic

\code{30yavg}: 30 year summary statistic
}
\keyword{datasets}
