% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_apireq_methods.R
\name{plot.ca_apireq}
\alias{plot.ca_apireq}
\title{Plot a ca_apireq object}
\usage{
\method{plot}{ca_apireq}(x, basemap = "Esri.NatGeoWorldMap", ...)
}
\arguments{
\item{x}{Cal-Adapt API request}

\item{basemap}{The name of a basemap tile layer (see tm_basemap)}

\item{...}{Unused}
}
\description{
Plot a ca_apireq object
}
