% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_apireq_methods.R
\name{print.ca_apireq}
\alias{print.ca_apireq}
\title{Print a ca_apireq object}
\usage{
\method{print}{ca_apireq}(x)
}
\arguments{
\item{x}{Cal-Adapt API request}
}
\description{
Print a ca_apireq object
}
