% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_db_info.R
\name{print.ca_db_info}
\alias{print.ca_db_info}
\title{Print a ca_db_info object}
\usage{
\method{print}{ca_db_info}(x, ...)
}
\arguments{
\item{x}{An object of class ca_db_info}

\item{...}{Unused}
}
\description{
Print a ca_db_info object
}
