% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_catalog_rs.R
\name{ca_catalog_rs}
\alias{ca_catalog_rs}
\title{Get the Cal-Adapt raster series data catalog}
\usage{
ca_catalog_rs(quiet = FALSE)
}
\arguments{
\item{quiet}{Suppress messages}
}
\value{
A tibble with columns of information about the raster series available through the Cal-Adapt API.
}
\description{
Get a local copy of the Cal-Adapt raster series data catalog
}
\details{
This retrieves a local copy of the Cal-Adapt 'catalog' of raster series available through the Cal-Adapt API.
A copy of the catalog comes with caladaptR. You can also fetch a new copy using \link{ca_catalog_fetch}.
}
\seealso{
\code{\link{ca_catalog_fetch}}, \code{\link{ca_getcache}}
}
