% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_example_apireq.R
\name{ca_example_apireq}
\alias{ca_example_apireq}
\title{Example API requests}
\usage{
ca_example_apireq(n)
}
\arguments{
\item{n}{Which sample API request to return (1..5)}
}
\description{
Example API requests
}
\details{
These sample API requests can be used in demos, documentation, and tests.

n = 1: Basic API request - one point, two CGMs, 20 years of annual data.
n = 2: Three Congressional districts, monthly data, 4 years
n = 3: sf data frame with one feature, 1 GCM, 1 scenario, 2 years of daily data
n = 4: Simple feature data frame with two multipolygons, 1 GCM, 1 scenario, 20 years of annual data
}
