% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_cache.R
\name{ca_getcache}
\alias{ca_getcache}
\alias{ca_setcache}
\title{Manage Cache Directory}
\usage{
ca_getcache(default = TRUE, quiet = TRUE)

ca_setcache(cache_dir, make_dir = FALSE, write = TRUE, quiet = FALSE)
}
\arguments{
\item{default}{Use a default cache directory if no other has been set}

\item{quiet}{Suppress messages}

\item{cache_dir}{The directory for cached data, can also be 'default' to use the default location}

\item{make_dir}{Make the directory if needed, logical}

\item{write}{Save the cache directory in the \code{.Renviron} file for persistence across R sessions}
}
\description{
View and set the directory for the data catalog
}
\details{
\code{caladaptr} has the ability to store copies of objects downloaded
from Cal-Adapt. An example of this would be the raster series data catalog and the
geometries of AOI Presets.

NOTE: \code{caladaptr} does \emph{not} cache climate data fetched from Cal-Adapt. Every time
you call a function that fetches data (e.g., \code{\link{ca_getvals}}), data is retrieved fresh.

If \code{default = TRUE}, a default cache directory will be used (\link[rappdirs]{user_cache_dir})
\emph{if} one hasn't already been set with \code{ca_setcache}.
}
\section{Functions}{
\itemize{
\item \code{ca_setcache}: Set cache directory
}}

\seealso{
\code{\link{ca_catalog_rs}}
}
