% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_cache.R
\name{ca_getcache}
\alias{ca_getcache}
\alias{ca_setcache}
\title{Manage Cache Directory}
\usage{
ca_getcache(default = TRUE, quiet = TRUE)

ca_setcache(dir = "~/.R/caladaptr", write = TRUE, quiet = FALSE)
}
\arguments{
\item{default}{Use a default cache directory if no other has been set}

\item{quiet}{Suppress messages}

\item{dir}{The directory for  cached EXIF data (must exist)}

\item{write}{Write directory location to .Renviron}
}
\description{
View and set the directory for the data catalog
}
\details{
\code{caladaptr} stores a local copy of the Cal-Adapt data catalog.
The data catalog can be refreshed by running \code{\link{ca_catalog_update}}.

Cached results are saved as native R objects.

If \code{default = TRUE}, a default directory for the cache (\emph{~/.R/caladaptr})
will be used if another one has not already been set.
}
\section{Functions}{
\itemize{
\item \code{ca_setcache}: Set cache directory
}}

\seealso{
\code{\link{ca_catalog_update}}
}
