% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_cache.R
\name{ca_getcache}
\alias{ca_getcache}
\alias{ca_setcache}
\title{Manage Cache Directory}
\usage{
ca_getcache(default = TRUE, quiet = TRUE)

ca_setcache(cache_dir, make_dir = FALSE, write = TRUE, quiet = FALSE)
}
\arguments{
\item{default}{Use a default cache directory if no other has been set}

\item{quiet}{Suppress messages}

\item{cache_dir}{The directory for cached data, can also be 'default' to use the default location}

\item{make_dir}{Make the directory if needed, logical}

\item{write}{Save the cache directory in the \code{.Renviron} file for persistence across R sessions}
}
\description{
View and set the directory for the data catalog
}
\details{
\code{caladaptr} has the ability to store copies of objects downloaded
from Cal-Adapt. An example of this would be the raster series data catalog.

If \code{default = TRUE}, a default directory for the cache (\emph{~/.R/caladaptr})
will be used if another one has not already been set.
}
\section{Functions}{
\itemize{
\item \code{ca_setcache}: Set cache directory
}}

\seealso{
\code{\link{ca_catalog_rs}}
}
