% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_cache.R
\name{ca_getcache}
\alias{ca_getcache}
\alias{ca_setcache}
\title{Manage cache directory}
\usage{
ca_getcache(quiet = TRUE)

ca_setcache(
  cache_dir = NULL,
  make_dir = TRUE,
  save = TRUE,
  reset = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{quiet}{Suppress messages}

\item{cache_dir}{The directory for cached data}

\item{make_dir}{Make the directory if needed, logical}

\item{save}{Save the cache directory in the \code{.Renviron} file for persistence across R sessions}

\item{reset}{Change to the default location}
}
\description{
View and set the directory for the data catalog
}
\details{
\code{caladaptr} has the ability to store copies of objects downloaded
from Cal-Adapt. An example of this would be the raster series data catalog and the
geometries of AOI Presets.

NOTE: In general \code{caladaptr} does \emph{not} cache climate data fetched from Cal-Adapt. Every time
you call a function that fetches data (e.g., \code{\link{ca_getvals_tbl}}), data is retrieved fresh.
The exception to this is \code{\link{ca_getvals_db}}, which has arguments you can pass to cache retrieved
values into a local SQLite database to explicitly avoid downloading data twice.

The default location for the cache directory is given by \code{R_user_dir("caladaptr", "cache")}.
A custom location can be set with \code{ca_setcache}.
}
\section{Functions}{
\itemize{
\item \code{ca_setcache}: Set cache directory
}}

\seealso{
\code{\link{ca_catalog_rs}}, \code{\link{ca_aoipreset_geom}}, \code{\link{ca_locagrid_geom}}
}
