% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_getrst.R
\name{ca_getrst}
\alias{ca_getrst}
\title{Get raster from an API request object}
\usage{
ca_getrst(slug, quiet = TRUE)
}
\arguments{
\item{slug}{A Cal-Adapt API slug}

\item{quiet}{Suppress messages}
}
\description{
Get raster from an API request object
}
\seealso{
\code{\link{ca_catalog_rs}}
}
