% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_getrst_stars.R
\name{ca_getrst_stars}
\alias{ca_getrst_stars}
\title{Get values from an API request object as a tibble}
\usage{
ca_getrst_stars(
  x,
  out_dir = NULL,
  merge_geoms = TRUE,
  sidecar_write = TRUE,
  stop_on_err = TRUE,
  quiet = FALSE,
  debug = FALSE
)
}
\arguments{
\item{x}{A Cal-Adapt API request}

\item{out_dir}{Where the output tif files should be written}

\item{sidecar_write}{Save a small sidecar file with the TIF file containing additional attribute info}

\item{stop_on_err}{Stop if the server returns an error}

\item{quiet}{Suppress messages}

\item{debug}{Print additional output at the console}
}
\value{
A tibble
}
\description{
Get values from an API request object as a tibble
}
\details{
This will download a multi-layer raster for your study area, convert it to a stars object, and export it to
TIF files. Note this will only work for areas-of-interest. If you want to download rasters for a large area (e.g., the
whole state) you're better off downloading NetCDF files from the Cal-Adapt \href{http://albers.cnr.berkeley.edu/data/}{data server}.

If \code{sidecar_write = TRUE}, a small file with the same base name as the TIF will be saved. This sidecar file contains
attributes of a space-time-array not perserved by TIFs. You can import the TIF file back into R as a stars object
with \code{\link{ca_read_stars}}.
}
