% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_getrst_stars.R
\name{ca_getrst_stars}
\alias{ca_getrst_stars}
\title{Get cropped rasters}
\usage{
ca_getrst_stars(
  x,
  out_dir = NULL,
  mask = TRUE,
  merge_geoms = FALSE,
  sidecar = TRUE,
  stop_on_err = TRUE,
  overwrite = FALSE,
  normalize_path = FALSE,
  debug = FALSE,
  quiet = FALSE,
  write_sidecar = deprecated()
)
}
\arguments{
\item{x}{A Cal-Adapt API request}

\item{out_dir}{Where the output TIF files should be written}

\item{mask}{Mask pixels outside the location of interest with NA values}

\item{merge_geoms}{Whether to merge geometries, see Details}

\item{sidecar}{Save a small sidecar file with the TIF file containing additional attribute info}

\item{stop_on_err}{Stop if the server returns an error}

\item{overwrite}{Re-download and overwrite existing files}

\item{normalize_path}{Expand and normalize output file names}

\item{debug}{Print additional output at the console}

\item{quiet}{Suppress messages}

\item{write_sidecar}{Deprecated}
}
\value{
A vector of TIF file names. If \code{normalize_path = TRUE} the output file names will be expanded (absolute) and use standard slashes for the OS (see normalizePath).
}
\description{
Download a cropped raster for an API request
}
\details{
This will download time series cropped raster(s) for your study area, convert them to stars objects, and export them
as tif files. If \code{mask = TRUE}, pixels values outside the area of interest will be set to \code{NA} (\code{mask} is ignored
for point locations). To get a single raster per dataset that encompasses all the locations,
pass \code{merge_geoms = TRUE}.

Note this will only work for areas-of-interest small enough for the Cal-Adapt API to handle (i.e., smaller than
San Bernadino County). If you want to download
rasters for a large area (e.g., the whole state of California) you're better off downloading NetCDF files from the
Cal-Adapt \href{http://albers.cnr.berkeley.edu/data/}{data server}.

If \code{sidecar = TRUE}, a small file with the same base name as the tif will be saved. This sidecar file contains
attributes of a space-time-array not preserved by tifs. You can import the tif file back into R as a stars object
with \code{\link{ca_read_stars}}.

This function merely downloads the cropped rasters to disk and returns the filenames. To work with cropped rasters
as stars objects within R, import them using \code{\link{ca_read_stars}}. You can also import the TIF files with other
packages or software.
}
\seealso{
\code{\link{ca_read_stars}}
}
