% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_getvals.R
\name{ca_getvals}
\alias{ca_getvals}
\title{Get values from an API request object}
\usage{
ca_getvals(x, quiet = FALSE, debug = FALSE, use_events = TRUE)
}
\arguments{
\item{x}{A Cal-Adapt API request}

\item{quiet}{Suppress messages}

\item{debug}{Print additional output at the console}

\item{use_events}{Use the events end point, logical}
}
\description{
Get values from an API request object
}
