% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_getvals_tbl.R
\name{ca_getvals_tbl}
\alias{ca_getvals_tbl}
\title{Get values from an API request object as a tibble}
\usage{
ca_getvals_tbl(x, quiet = FALSE, debug = FALSE, stop_on_err = TRUE)
}
\arguments{
\item{x}{A Cal-Adapt API request}

\item{quiet}{Suppress messages}

\item{debug}{Print additional output at the console}

\item{stop_on_err}{Stop if the server returns an error}
}
\value{
A tibble
}
\description{
Get values from an API request object as a tibble
}
