% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_loc_pt.R
\name{ca_loc_pt}
\alias{ca_loc_pt}
\title{Add point location(s) to a Cal-Adapt API request}
\usage{
ca_loc_pt(x = ca_apireq(), coords, id = NULL)
}
\arguments{
\item{x}{A Cal-Adapt API request}

\item{coords}{A two-column matrix or data frame}

\item{id}{Unique id values for each point}
}
\description{
Specifies point location(s) a Cal-Adapt API call should retrieve
}
\details{
\code{coords} should be a two-column matrix or data frame with the first column
containing the x (longitude) values of the points of interest, and the second column
containing the y (latitude) values. Projected coordinates can not be used with this
function (but see \code{\link{ca_loc_sf}}).

\code{id} should be vector that uniquely identify the points. If omitted, row numbers will be used.
}
\seealso{
\code{\link{ca_loc_sf}}, \code{\link{ca_apireq}}
}
