% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_options.R
\name{ca_options}
\alias{ca_options}
\title{Add processing options to a Cal-Adapt API call}
\usage{
ca_options(
  x = ca_apireq(),
  spatial_ag = c("none", "mean", "max", "median", "min", "sum")[1],
  temporal_ag = NA
)
}
\arguments{
\item{x}{Cal-Adapt API request}

\item{spatial_ag}{Spatial aggregation function for polygon locs.}

\item{temporal_ag}{List object specifying unit of time and summary function(s) for temporal aggregation. NOT YET SUPPORTED}
}
\description{
Specify processing options for a Cal-Adapt API call
}
\details{

}
