% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_period.R
\docType{data}
\name{ca_period}
\alias{ca_period}
\title{Temporal aggregation periods}
\format{
An character vector with three names of temporal aggregation periods
}
\source{
\url{https://berkeley-gif.github.io/caladapt-docs/data-catalog.html#period}
}
\usage{
data(ca_period)
}
\description{
Temporal aggregation periods for raster series
}
\details{
The following temporal aggregation periods for raster series are available thru the Cal-Adapt API.

\code{year}: Annual averages

\code{month}: Monthly averages

\code{day}: Daily values
}
\keyword{datasets}
