% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_read_stars.R
\name{ca_read_stars}
\alias{ca_read_stars}
\title{Read a stars object from disk}
\usage{
ca_read_stars(x, read_sidecar = TRUE)
}
\arguments{
\item{x}{A single file name}

\item{read_sidecar}{Whether to read the sidecar file if it exists, logical}
}
\description{
Read a stars object from disk
}
\details{
This function can be used to read a tif file downloaded by \code{\link{ca_getrst_stars}}. It
is a lightweight wrapper around stars::read_stars(), with the ability to create a sidecar
file (also created by \code{\link{ca_getrst_stars}}) which restores all of the attributes of
the array dimensions that are not preserved by the tif format.
}
\seealso{
\code{\link{ca_getrst_stars}}
}
