% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_scenario.R
\name{ca_scenario}
\alias{ca_scenario}
\title{Assign emission scenario(s) to a Cal-Adapt API call}
\usage{
ca_scenario(x = ca_apireq(), scenario)
}
\arguments{
\item{x}{Cal-Adapt API request}

\item{scenario}{Abbreviation of emissions scenario(s)}
}
\description{
Specifies emission scenario(s) for retrieval
}
\details{
For valid options for \code{scenario}, see \code{\link{scenarios}}.
}
