% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_vals2tbl.R
\name{ca_vals2tbl}
\alias{ca_vals2tbl}
\title{Converts a Cal-Adapt values list to a data frame}
\usage{
ca_vals2tbl(
  x,
  feat_id = TRUE,
  dt = TRUE,
  cvar = TRUE,
  period = TRUE,
  gcm = TRUE,
  scenario = TRUE,
  spatial_ag = TRUE
)
}
\arguments{
\item{x}{A Cal-Adapt Values Object}

\item{feat_id}{Include feature id as a column}

\item{dt}{Include date}

\item{cvar}{Include climate variable}

\item{period}{Include period}

\item{gcm}{Include gcm}

\item{scenario}{Include scenario}

\item{spatial_ag}{Include spatial aggregation function}
}
\description{
Converts a Cal-Adapt values list to a data frame
}
\seealso{
\code{\link{ca_getvals}}
}
