% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_apireq_methods.R
\name{plot.ca_apireq}
\alias{plot.ca_apireq}
\title{Plot a ca_apireq object}
\usage{
\method{plot}{ca_apireq}(
  x,
  basemap = c("Esri.NatGeoWorldMap", "OpenStreetMap")[1],
  locagrid = FALSE,
  static = FALSE,
  ...
)
}
\arguments{
\item{x}{Cal-Adapt API request}

\item{basemap}{The name of a basemap tile layer (see tm_basemap)}

\item{locagrid}{Overlay a portion of the LOCA downscaled grid}

\item{static}{Plot a static map instead of a interactive leaflet map}

\item{...}{Unused}
}
\description{
Plot a ca_apireq object
}
